Optimize a bit, add README
In main(), an array of structures is declared, each worker thread is passed
one element of that array, containing the specification of the work to be done
and a "result" variable to store the result.

Initially each thread was using "result" from the passed struct during its
sum of squares computations, reading and updating it every iteration of the
computation loop. This seemed to create a bottleneck, such that it was taking
about as long for two threads to do the same work vs a single thread (things
got a bit better with more worker threads).

The bottleneck above was relieved by having each thread allocate a local "result"
var to use when doing computations. With this change, it takes about half the time
with two worker threads vs one.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>